Fix bug that prevents packages from installing when runAsUser=true (#95)
The bug's root cause was the addition of the Version parameter to enable specifying package versions. The implementation adds an empty string "" whenever the version isn't specified, and this empty string in the command line that calls winget.exe causes it to error out.

The fix is to clean the command line so it doesn't contain empty string "" when involking the task with an empty version parameter.