Update git-clone task to align with winget task on scheduled work (#96)
This addresses a potential issue when a git-clone operation happens ahead of a winget operation during a customization flow. The contents of runAsUser should be the same for both scenarios so it won't matter what ran first, the scheduled task will have the same result.

Also, aligning the PS7 and Winget installation code in main.ps1 to the levels of the winget task.