Delete dangling target directory ahead of running git clone (#98)
When a previous git clone operation has failed, the target directory will exist in the machine but remain empty. When attempting to clone again, the existing directory will signal git.exe to fail. In order to not run into this situation, we'll try to ensure that the target directory doesn't exist ahead of running git clone operations.